TPM: Fixed misinterpretation of tpm_transmit() parameter
The last parameter of tpm_transmit() reflects the amount
of data expected from the device, and not the buffer size
being supplied to it. At the specific case of tpm_write(),
given the command is coming from userspace, such information
can't be parsed directly from such buffer, so we just set
it to the maximum the input buffer can handle and let the
userspace API decide whether it got the expected number
of bytes.

This change partially reverts the one introduced by
commit adfea973d

Signed-off-by: Rajiv Andrade <srajiv@linux.vnet.ibm.com>