TPM: NSC driver X86 dependency fix
The tpm_nsc driver is supposed to support old 1.1 NSC chips,
that are only found on x86 based machines. A previous commit
removed its PNP dependency, that in fact wasn't necessary, but
also allowed it be built for other architectures not supported
by it. This then caused kernel oops on PPC based machines. I'm
placing a x86 dependency back correctly.

Signed-off-by: Rajiv Andrade <srajiv@linux.vnet.ibm.com>